Moved BaseDesign temporary to keep distribution folder clean
